refactored models according to latest branches for easier merging
