+++ src/etools/applications/field_monitoring/fm_settings/models/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+
+from etools.applications.field_monitoring.fm_settings.models.cp_outputs import CPOutputConfig
+from etools.applications.field_monitoring.fm_settings.models.locations import LocationSite
+from etools.applications.field_monitoring.fm_settings.models.methods import FMMethodType
