+++ src/etools/applications/field_monitoring/fm_settings/tests/test_views.py
@@ -8,9 +8,9 @@ from unicef_locations.tests.factories import LocationFactory
 
 from etools.applications.EquiTrack.tests.cases import BaseTenantTestCase
 from etools.applications.field_monitoring.fm_settings.models import CPOutputConfig
-from etools.applications.field_monitoring.fm_settings.tests.base import FMBaseTestCaseMixin
 from etools.applications.field_monitoring.fm_settings.tests.factories import FMMethodTypeFactory, LocationSiteFactory, \
     CPOutputConfigFactory, FMMethodFactory
+from etools.applications.field_monitoring.tests.base import FMBaseTestCaseMixin
 from etools.applications.partners.models import PartnerType
 from etools.applications.partners.tests.factories import PartnerFactory
 from etools.applications.reports.models import ResultType
