+++ src/etools/applications/field_monitoring/groups.py
@@ -1,0 +1,5 @@
+from etools.libraries.djangolib.models import GroupWrapper
+
+
+UNICEFUser = GroupWrapper(code='unicef_user',
+                          name='UNICEF User')
