+++ src/etools/applications/field_monitoring/tests/base.py
