+++ src/etools/applications/field_monitoring/views.py
@@ -1,0 +1,15 @@
+from rest_framework.permissions import IsAuthenticated
+
+from unicef_restlib.pagination import DynamicPageNumberPagination
+from unicef_restlib.views import SafeTenantViewSetMixin, MultiSerializerViewSetMixin
+
+from etools.applications.permissions2.metadata import BaseMetadata
+
+
+class FMBaseViewSet(
+    SafeTenantViewSetMixin,
+    MultiSerializerViewSetMixin,
+):
+    metadata_class = BaseMetadata
+    pagination_class = DynamicPageNumberPagination
+    permission_classes = [IsAuthenticated, ]
